backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-dotenv
python-jose
passlib[bcrypt]
pytest
httpx